// src/_proto/Protos/bookmarks.ts
export interface BookmarkItem {
    /**
     * @generated from protobuf field: int32 novelId = 1;
     */
    novelId: number;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ChapterItem chapter = 2;
     */
    chapter?: ChapterItem;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp readAt = 3;
     */
    readAt?: Timestamp;
    /**
     * @generated from protobuf field: bool locked = 4;
     */
    locked: boolean;
    /**
     * @generated from protobuf field: bool notificationsEnabled = 5;
     */
    notificationsEnabled: boolean;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ChapterItem latestChapter = 6;
     */
    latestChapter?: ChapterItem;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ReadingListType listType = 7;
     */
    listType: ReadingListType;
}
/**
 * @generated from protobuf enum wuxiaworld.api.v2.ReadingListType
 */
export enum ReadingListType {
    /**
     * @generated from protobuf enum value: Current = 0;
     */
    Current = 0,
    /**
     * @generated from protobuf enum value: Completed = 1;
     */
    Completed = 1,
    /**
     * @generated from protobuf enum value: Dropped = 3;
     */
    Dropped = 3,
    /**
     * @generated from protobuf enum value: OnHold = 4;
     */
    OnHold = 4,
    /**
     * @generated from protobuf enum value: Manual = 5;
     */
    Manual = 5
}
// ...
class BookmarkItem$Type extends MessageType<BookmarkItem> {
    constructor() {
        super("wuxiaworld.api.v2.BookmarkItem", [
            { no: 1, name: "novelId", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "chapter", kind: "message", T: () => ChapterItem },
            { no: 3, name: "readAt", kind: "message", T: () => Timestamp },
            { no: 4, name: "locked", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 5, name: "notificationsEnabled", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 6, name: "latestChapter", kind: "message", T: () => ChapterItem },
            { no: 7, name: "listType", kind: "enum", T: () => ["wuxiaworld.api.v2.ReadingListType", ReadingListType] }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.BookmarkItem
 */
export const BookmarkItem = new BookmarkItem$Type();