// src/_proto/Protos/novels.ts
export interface NovelChapterInfo {
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ChapterItem firstChapter = 1;
     */
    firstChapter?: ChapterItem;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ChapterItem latestChapter = 2;
     */
    latestChapter?: ChapterItem;
    /**
     * @generated from protobuf field: google.protobuf.Int32Value chapterCount = 3;
     */
    chapterCount?: Int32Value;
    /**
     * @generated from protobuf field: repeated wuxiaworld.api.v2.ChapterGroupItem chapterGroups = 4;
     */
    chapterGroups: ChapterGroupItem[];
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.NovelKarmaInfo
 */
export interface NovelKarmaInfo {
    /**
     * @generated from protobuf field: bool isActive = 1;
     */
    isActive: boolean;
    /**
     * @generated from protobuf field: bool isFree = 2;
     */
    isFree: boolean;
    /**
     * @generated from protobuf field: customTypes.DecimalValue maxFreeChapter = 3;
     */
    maxFreeChapter?: DecimalValue;
    /**
     * @generated from protobuf field: bool canUnlockWithVip = 4;
     */
    canUnlockWithVip: boolean;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.NovelReviewInfo
 */
export interface NovelReviewInfo {
    /**
     * @generated from protobuf field: int32 count = 1;
     */
    count: number;
    /**
     * @generated from protobuf field: google.protobuf.DoubleValue rating = 2;
     */
    rating?: DoubleValue;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.NovelSponsorInfo
 */
export interface NovelSponsorInfo {
    /**
     * @generated from protobuf field: google.protobuf.Int32Value advanceChapterCount = 1;
     */
    advanceChapterCount?: Int32Value;
    /**
     * @generated from protobuf field: google.protobuf.StringValue description = 2;
     */
    description?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.BoolValue hasAnyPlans = 3;
     */
    hasAnyPlans?: BoolValue;
    /**
     * @generated from protobuf field: repeated wuxiaworld.api.v2.SponsorPlanItem plans = 4;
     */
    plans: SponsorPlanItem[];
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.NovelItem
 */
export interface NovelItem {
    /**
     * @generated from protobuf field: int32 id = 1;
     */
    id: number;
    /**
     * @generated from protobuf field: string name = 2;
     */
    name: string;
    /**
     * @generated from protobuf field: string slug = 3;
     */
    slug: string;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NovelItem.Status status = 4;
     */
    status: NovelItem_Status;
    /**
     * @generated from protobuf field: string abbreviation = 5;
     */
    abbreviation: string;
    /**
     * @generated from protobuf field: google.protobuf.StringValue language = 6;
     */
    language?: StringValue;
    /**
     * @generated from protobuf field: bool visible = 7;
     */
    visible: boolean;
    /**
     * @generated from protobuf field: google.protobuf.StringValue description = 8;
     */
    description?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.StringValue synopsis = 9;
     */
    synopsis?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.StringValue coverUrl = 10;
     */
    coverUrl?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.StringValue translatorName = 11;
     */
    translatorName?: StringValue;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.UserItem translator = 12;
     */
    translator?: UserItem;
    /**
     * @generated from protobuf field: google.protobuf.StringValue authorName = 13;
     */
    authorName?: StringValue;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NovelKarmaInfo karmaInfo = 14;
     */
    karmaInfo?: NovelKarmaInfo;
    /**
     * @generated from protobuf field: repeated string tags = 15;
     */
    tags: string[];
    /**
     * @generated from protobuf field: repeated string genres = 16;
     */
    genres: string[];
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NovelReviewInfo reviewInfo = 17;
     */
    reviewInfo?: NovelReviewInfo;
    /**
     * @generated from protobuf field: bool isSneakPeek = 18;
     */
    isSneakPeek: boolean;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NovelSponsorInfo sponsorInfo = 19;
     */
    sponsorInfo?: NovelSponsorInfo;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp createdAt = 20;
     */
    createdAt?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.StringValue coverBlurHash = 21;
     */
    coverBlurHash?: StringValue;
    /**
     * @generated from protobuf field: bool isFeatured = 22;
     */
    isFeatured: boolean;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.NovelChapterInfo chapterInfo = 23;
     */
    chapterInfo?: NovelChapterInfo;
    /**
     * @generated from protobuf field: google.protobuf.StringValue licensedFrom = 24;
     */
    licensedFrom?: StringValue;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.SeriesItem series = 25;
     */
    series?: SeriesItem;
    /**
     * @generated from protobuf field: bool isTextToSpeechEnabled = 26;
     */
    isTextToSpeechEnabled: boolean;
}
/**
 * @generated from protobuf enum wuxiaworld.api.v2.NovelItem.Status
 */
export enum NovelItem_Status {
    /**
     * @generated from protobuf enum value: Finished = 0;
     */
    Finished = 0,
    /**
     * @generated from protobuf enum value: Active = 1;
     */
    Active = 1,
    /**
     * @generated from protobuf enum value: Hiatus = 2;
     */
    Hiatus = 2,
    /**
     * @generated from protobuf enum value: All = -1;
     */
    All = -1
}
// ...
class NovelChapterInfo$Type extends MessageType<NovelChapterInfo> {
    constructor() {
        super("wuxiaworld.api.v2.NovelChapterInfo", [
            { no: 1, name: "firstChapter", kind: "message", T: () => ChapterItem },
            { no: 2, name: "latestChapter", kind: "message", T: () => ChapterItem },
            { no: 3, name: "chapterCount", kind: "message", T: () => Int32Value },
            { no: 4, name: "chapterGroups", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => ChapterGroupItem }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.NovelChapterInfo
 */
export const NovelChapterInfo = new NovelChapterInfo$Type();
// @generated message type with reflection information, may provide speed optimized methods
class NovelKarmaInfo$Type extends MessageType<NovelKarmaInfo> {
    constructor() {
        super("wuxiaworld.api.v2.NovelKarmaInfo", [
            { no: 1, name: "isActive", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 2, name: "isFree", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 3, name: "maxFreeChapter", kind: "message", T: () => DecimalValue },
            { no: 4, name: "canUnlockWithVip", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.NovelKarmaInfo
 */
export const NovelKarmaInfo = new NovelKarmaInfo$Type();
// @generated message type with reflection information, may provide speed optimized methods
class NovelReviewInfo$Type extends MessageType<NovelReviewInfo> {
    constructor() {
        super("wuxiaworld.api.v2.NovelReviewInfo", [
            { no: 1, name: "count", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "rating", kind: "message", T: () => DoubleValue }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.NovelReviewInfo
 */
export const NovelReviewInfo = new NovelReviewInfo$Type();
// @generated message type with reflection information, may provide speed optimized methods
class NovelSponsorInfo$Type extends MessageType<NovelSponsorInfo> {
    constructor() {
        super("wuxiaworld.api.v2.NovelSponsorInfo", [
            { no: 1, name: "advanceChapterCount", kind: "message", T: () => Int32Value },
            { no: 2, name: "description", kind: "message", T: () => StringValue },
            { no: 3, name: "hasAnyPlans", kind: "message", T: () => BoolValue },
            { no: 4, name: "plans", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => SponsorPlanItem }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.NovelSponsorInfo
 */
export const NovelSponsorInfo = new NovelSponsorInfo$Type();
// @generated message type with reflection information, may provide speed optimized methods
class NovelItem$Type extends MessageType<NovelItem> {
    constructor() {
        super("wuxiaworld.api.v2.NovelItem", [
            { no: 1, name: "id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "slug", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "status", kind: "enum", T: () => ["wuxiaworld.api.v2.NovelItem.Status", NovelItem_Status] },
            { no: 5, name: "abbreviation", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 6, name: "language", kind: "message", T: () => StringValue },
            { no: 7, name: "visible", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 8, name: "description", kind: "message", T: () => StringValue },
            { no: 9, name: "synopsis", kind: "message", T: () => StringValue },
            { no: 10, name: "coverUrl", kind: "message", T: () => StringValue },
            { no: 11, name: "translatorName", kind: "message", T: () => StringValue },
            { no: 12, name: "translator", kind: "message", T: () => UserItem },
            { no: 13, name: "authorName", kind: "message", T: () => StringValue },
            { no: 14, name: "karmaInfo", kind: "message", T: () => NovelKarmaInfo },
            { no: 15, name: "tags", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 16, name: "genres", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 17, name: "reviewInfo", kind: "message", T: () => NovelReviewInfo },
            { no: 18, name: "isSneakPeek", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 19, name: "sponsorInfo", kind: "message", T: () => NovelSponsorInfo },
            { no: 20, name: "createdAt", kind: "message", T: () => Timestamp },
            { no: 21, name: "coverBlurHash", kind: "message", T: () => StringValue },
            { no: 22, name: "isFeatured", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 23, name: "chapterInfo", kind: "message", T: () => NovelChapterInfo },
            { no: 24, name: "licensedFrom", kind: "message", T: () => StringValue },
            { no: 25, name: "series", kind: "message", T: () => SeriesItem },
            { no: 26, name: "isTextToSpeechEnabled", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.NovelItem
 */
export const NovelItem = new NovelItem$Type();