// src/_proto/Protos/pricing.ts
export interface KarmaPricingModel {
    /**
     * @generated from protobuf field: customTypes.DecimalValue freeChapters = 1;
     */
    freeChapters?: DecimalValue;
}
// ...
class KarmaPricingModel$Type extends MessageType<KarmaPricingModel> {
    constructor() {
        super("wuxiaworld.api.v2.KarmaPricingModel", [
            { no: 1, name: "freeChapters", kind: "message", T: () => DecimalValue }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.KarmaPricingModel
 */
export const KarmaPricingModel = new KarmaPricingModel$Type();